Improve performance when receiving a large number of added/removed files
This commit adds a separate implementation that is faster when adding/removing
a large number of files (in a synthetic benchamrk, the turning point is around
5000 items).

My real-world test case is ~5k new files and ~55k new evals taking almost two
minutes to be processed.

Each add()/remove() is O(myModel.size()), so the current implementation costs

    O(startSize * (startSize - 1) / 2) - (afterRemove * (afterRemove - 1) / 2)) + /* removal */
    O(afterAdd * (afterAdd - 1) / 2) - (afterRemove * (afterRemove - 1) / 2)) + /* addition */
    O(removed + added) /* one notification for each operation */

which, for small added/removed simplifies to:

    O((added + removed) * afterAdd)

for large added/removed simplifies to:

    O(startSize * startSize + afterAdd * afterAdd)

The new alternative implementation costs

    O(removed * log(startSize)) + /* removal */
    O(added) + /* addition */
    O(afterAdd) + O(afterAdd * log(afterAdd)) /* readding and sorting */
    O(afterAdd) /* a single notification for the whole range */

which (for most cases) simplifies to:

   O(afterAdd * log(afterAdd))

So the current version is good for "small" added/removed; where the exact value
needs to be determined with benchmarks.

(cherry picked from commit ce3fc66)