Simplified QuoteHandler
We can't actually detect if we have non-closed literal ahead after re-lexing.
Default quote handler works when quotes are part of literal psi elements and there can't be multi-line strings, but this is not our case.
Afraid it's not technically possible, we can't even calculate quotes ahead, because of regexps and so on.