Implemented blessed value support
FYI: this is a kind of support, because it's blesses reference, not target.
In real world there is a, let's say - variable, which is blessed. So other reference from this variable going to be blessed too. But not with us.
This, probably, should be handled in values or when values is built.
But this solution is good enough for now.