Double slash after closing brace is now treated as or-defined operator
Fixes #2304

Technically this fix is incorrect, because this lexing should be context dependent and should be lexed differently in operator and operand context. But our lexer is light and has no information about context.

This fix will break things like:
```
{
  say 'hi';
}
/matchregexp/ or return;
```
But this case is really strange and may be workarounded with explicit `m` prefix. When original bug can't be worked around without using some excessive syntax.