Got rid of Project.getBaseDir for the remote debugging
This is a bit a lie, because we still have a fallback to the Project.getBasePath().
User may now override the path in the UI. Empty path should behave the same as before.
In generic run configuration we do have a contextual file to run, but in remote - we don't.
We can't fallback to the empty path, because current directory of java process is the same for all projects.