#2502 Adjusted quick documentation computation for elements
Since 2022.1 IDE passes some fake editor when computing custom element and we are unable to resolve and substitute perl with pod anymore.
Now we are doing this on generation step.
Additionally, some perl elements are now passed as context elements to the generation method and we need to compute pod context for them once again.

Fixes #2502