#2539 #CAMELCADE-5913 Improved wsl mount point initialization
## Problem
We have a number of activities involving paths mapping. And to make it right, we need a windows fs mount point. It is initialized lazily and this make it absolutely random - we don't know in which context we will try to do this. Reading mount point involves execution of the external tool which is now covered with modal progress, but it is not allowed in the action update pass.

## Solution
Before we pre-initialized all wsl distributions on the application start, but seems that distributions list was incomplete for some reason. Probably, some of them were installed in the runtime. Another flaw of this approach is spinning up all distributions, even unused ones.

After this change we will pre-initialize mount root on project opening if wsl sdk is selected for it and on changing sdk via `PerlProjectManager`. This should handle all actions cases, because actions always involve sdk selected for the project.

We are using `postStartupActivity` because `backgroundPostStartupActivity` are delayed for 5 seconds.

#CAMELCADE-5913 Fixed
Fixes #2539