#2548 added a fallback to the guessProjectDir for deparsed files
Some IDEs based on the build systems, like Rider, CLion and so on, may have no conventional content roots. They are using files as those and building project content by themselves on information from the build system. In rider case it is a solution file.

guessProjectDir should solve the issue with error reporting, but proper solution for the problem is making some project-level internal cache directory used as an implicit external library.