#2558 #2539 properly handle issues with wsl paths mappings, increased timeout to 10 seconds
If for some reason IDE fails to compute mapping for wsl in 5 seconds, it currently throws an NPE. Changed this behavior to the falling back to local path. It won't work and we probably should implement smarter behavior here, like ExecutionException with proper error message, but this looks pretty rare for now.

Our API is not ready for propagating ExecutionException in a lot of places. Most API users are getLocalPath/getRemotePath

Fixes #2558