#2569 pass library arguments to the prove itself
Previously, in-project and external libraries were passed to the testing scripts via PassEnv. `prove` was expected to work by itself.
Now we pass these paths to the `prove` as well. Probaby, this should be configurable, e.g. we could have a checkbox or two for passing external/in project libraries to the `prove`.
But should be good enough for now.

Fixes #2569