[Qodana] disabled caching
Qodana cache is huge and used only by qodana. Presumably it contains the same stuff we have in gradle. Also, it used to grow twice between runs.
This overflows cache capacity of 10 GB very fast and evicting useful stuff.

As soon as we run Qodana in parallel now, we have time to run until last step is finished.